+++ packages/form-render/src/type.ts
@@ -401,7 +401,7 @@ export interface FRProps extends Omit<AntdFormProps, 'form'> {
    */
   methods?: Record<string, Function>;
   operateExtra?: React.ReactNode;
-  maxWidth?: string;
+  maxWidth?: number | string;
   footer?: boolean | (() => React.ReactNode) | Partial<ActionProps> ;
 }
 
@@ -425,6 +425,7 @@ export interface SearchProps<RecordType> extends Omit<FRProps, 'form'> {
   resetText?: string;
   onSearch?: (search: any) => any;
   afterSearch?: (params: any) => any;
+  onReset?: (form: any) => void;
   widgets?: any;
   form?: any;
   [key:string]: any